MorseDataFunktion ready

+++ src/main/java/de/tuberlin/tubit/gitlab/anton/rudacov/functions/AssignKeyFunction.java
@@ -4,16 +4,16 @@ import de.tuberlin.tubit.gitlab.anton.rudacov.data.DataPoint;
 import de.tuberlin.tubit.gitlab.anton.rudacov.data.KeyedDataPoint;
 import org.apache.flink.api.common.functions.MapFunction;
 
-public class AssignKeyFunction implements MapFunction<DataPoint<Double>, KeyedDataPoint<Double>> {
+public class AssignKeyFunction implements MapFunction<DataPoint<Integer>, KeyedDataPoint<Integer>> {
 
-  private String key;
+    private String key;
 
-  public AssignKeyFunction(String key) {
-    this.key = key;
-  }
+    public AssignKeyFunction(String key) {
+        this.key = key;
+    }
 
-  @Override
-  public KeyedDataPoint<Double> map(DataPoint<Double> dataPoint) throws Exception {
-    return dataPoint.withKey(key);
-  }
+    @Override
+    public KeyedDataPoint<Integer> map(DataPoint<Integer> dataPoint) throws Exception {
+        return dataPoint.withKey(key);
+    }
 }
